pyproject.toml
[project]
name = "markdown-mindmap-studio"
version = "0.1.0"
description = "Interactive mindmap editor using Markdown files as single source of truth"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "claude-agent-sdk>=0.1.14",
    "fastapi>=0.124.0",
    "markdown>=3.10",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "uvicorn>=0.38.0",
    "weasyprint>=67.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "pytest-playwright>=0.7.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
